requirements.txt
pyyaml
bluepy